+++ cypress/integration/Validation/index.spec.js
@@ -17,7 +17,7 @@ describe('Validation', () => {
     it('should add class danger if input is invalid', () => {
       cy.name('Toggle')
         .click()
-        .name('Input1')
+        .name('InputCV')
         .parent()
         .should(($element) => {
           expect($element).to.have.length(1);
@@ -47,11 +47,27 @@ describe('Validation', () => {
     it('submit should fail if input field is set as invalid', () => {
       cy.name('Toggle')
         .click()
-        .name('Submit')
+        .name('SubmitCV')
         .click()
-        .name('Message')
-        .should('have.text', 'Submit failed');
+        .get('@consoleLog')
+        .should('be.calledWith', 'Submit failed');
     });
+
+    it('Should scroll to invalid fields', () => {
+      cy.name('SubmitScroll')
+        .click()
+        .get('@consoleLog')
+        .should('be.calledWith', 'Submit failed')
+        .get('.filedrop-content')
+        .isAtTop()
+        .name('fileDropScroll')
+        .attachFile('example.json', { subjectType: 'drag-n-drop' })
+        .name('SubmitScroll')
+        .click()
+        .get('.dropzone-content')
+        .isAtTop()
+    });
+
     it('should validate onBlur', () => {
       cy.get('#validationRequiredBlur')
         .find('input')